Several changes to the gamemode system in general
- Moved classes that modify or disable behaviour in Infinity into io.github.derechtepilz.infinity.gamemode.modification
- Changed name of Separator interface to Serializer
- Get rid of the GameModeChangeListener
- Get rid of the GameModeChangeEvent
- Moves logic for switching gamemodes from the deleted classes to a new GamemodeSwitchHandler
