+++ src/main/kotlin/io/github/derechtepilz/infinity/gamemode/modification/ChatHandler.kt
@@ -16,21 +16,28 @@
  *  along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package io.github.derechtepilz.infinity.gamemode
+package io.github.derechtepilz.infinity.gamemode.modification
 
-import com.destroystokyo.paper.event.player.PlayerAdvancementCriterionGrantEvent
-import io.github.derechtepilz.infinity.gamemode.Gamemode
 import io.github.derechtepilz.infinity.gamemode.getGamemode
+import io.papermc.paper.event.player.AsyncChatEvent
+import org.bukkit.Bukkit
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 
-class AdvancementListener : Listener {
+@Suppress("OverrideOnly")
+class ChatHandler : Listener {
 
 	@EventHandler
-	fun onGainCriterion(event: PlayerAdvancementCriterionGrantEvent) {
+	fun onChat(event: AsyncChatEvent) {
 		val player = event.player
-		if (player.getGamemode() == Gamemode.INFINITY) {
-			event.isCancelled = true
+		event.isCancelled = true
+		val playerGamemode = player.getGamemode()
+		for (p in Bukkit.getOnlinePlayers()) {
+			if (p.getGamemode() != playerGamemode) {
+				continue
+			}
+			val message = event.renderer().render(player, player.displayName(), event.message(), p)
+			p.sendMessage(message)
 		}
 	}
